plot8

+++ plot.py
@@ -12,7 +12,7 @@ for row in myreader:
 	# print(row[1])
 	if(row[1] == 'train_loss'):
 		continue
-	array.append(int(row[1]))
+	array.append(float(row[1]))
 print(array)
 # plt.plot(array)
 # plt.savefig('loss.png')
